Revert to int

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
  *
  */
 
-void print_number(long n)
+void print_number(int n)
 {
 	if (n < 0)
 	{
